CORE-794 fix overly sensitive rebuild
When switching between branches, git changes the modification times of
files that change. This caused the TLS files to be rebuilt even when
content didn't change.  `diff` is now used to detect when the
openssl.cnf content would change and uses this info to initiate a full
rebuild of TLS files.